requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
faiss-cpu
beautifulsoup4
requests
transformers>=4.41.0
sentence-transformers>=2.6.1